feat:抽奖API接口实现

+++ big-market-study-infrastructure/src/main/java/org/example/infrastructure/persistent/po/StrategyAward.java
@@ -26,7 +26,7 @@ public class StrategyAward {
     /**奖品中奖概率*/
     private BigDecimal awardCountRate;
     /**排序*/
-    private Long sort;
+    private Integer sort;
     /**创建时间*/
     private Date createTime;
     /**更新时间*/
